requirements.txt
Flask==3.1.0
PyYAML==6.0.2
python-dotenv==1.0.1
python-dateutil==2.9.0
Markdown==3.7
